Read the description of this commit!
I have finished part of the SRS, but I expect you guys to fill out user
specifications for each one of your modules. This is because it'l
honestly be a lot faster for you guys than it will be for me, since I
don't know your modules as well as you guys do. Look at the user
characteristics section in the overall description section to see what I
wrote for mine- it should really only take 10 minutes.