Added some revisions
What I see in the document is not at all similar to what I got in my
hands, especially for Xavier's place. Xavier, I think you didn't commit
your part of the UM properly. But in any case, just add a period at the
end of the sentence you have at 3.2.2.3 and it's all good