Added the base skeleton of CTCMain
Has all of the GUI code + some handling code. Need to add a LOT of handling code and also a way for changes that the private event handlers within the GUI class make to be accessible to my main class. (But I want to keep them private because I feel like they should be private.)