Attempting to use new linked list to get distance
Go to line 605 of CTCGUI to look at my approach to traversing the linked
list. It is currently getting a nullpointer exception.

(Summary - I am testing the green line - I start out at block 152, the
block closest to the yard, and I keep taking the next block until I get
to the block that I am looking for. I have only tried inputting block
numbers between 62 and 76 as destinations, as that is the simplest test
case.)