Added additional green track traversal
And just as important, improved the style and made it so that the
trainID will not be updated unless the train actually shows up on the
yard. This makes CTCTrainManager much less vulnerable to errors whenever
the wayside controller rejects a suggestion.
Also added a block tracker that will show the user a failure status and whether the block is occupied.